% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbskim.R
\name{dbskim}
\alias{dbskim}
\title{Skim a database table, getting usefulsummary statistics}
\usage{
dbskim(table_name, conn)
}
\arguments{
\item{table_name}{Database table name #TODO: support schemas}

\item{conn}{Database connection}
}
\value{
tibble of table summary statistics
}
\description{
`dbskim()` is inspired by \link[skimr]{skim}, quickly providing an overview of all the column names, types, and summary stats of a database table.
}
\examples{
conn <- DBI::dbConnect(RSQLite::SQLite(), ":memory:")
DBI::dbWriteTable(conn, "storms", dplyr::storms)
dbskim("storms", conn)
DBI::dbDisconnect(conn)

}
